data_pipeline/requirements.txt
# 数据采集相关
paho-mqtt==1.6.1
websockets==12.0
aiohttp==3.8.6

# 数据处理相关
pyspark==3.5.0
py4j==0.10.9.7

# 数据存储相关
redis==5.0.1
influxdb-client==1.39.0
sqlalchemy==2.0.25

# 消息队列
kafka-python==2.0.2

# 调度系统
# 注意：apache-airflow安装复杂，已移除。项目使用内置的scheduler.py实现调度功能

# 配置管理
pyyaml==6.0.1
python-dotenv==1.0.0

# 监控相关
prometheus-client==0.19.0

# 工具库
# 注意：logging、datetime、uuid是Python标准库，不需要通过pip安装

# 机器学习
tensorflow==2.15.0
tensorflow-serving-api==2.15.0
scikit-learn==1.4.0
librosa==0.10.1
xgboost==2.0.0
joblib==1.3.0
numpy==1.26.0
httpx==0.26.0
pandas==2.1.0

# 容器化
docker==6.1.3
# WebSocket支持
Flask-SocketIO==5.3.4
python-socketio==5.10.0
python-engineio==4.8.0
